requirements.txt
flask
raven
requests
python-dotenv
mysqlclient
uwsgi==2.0.16
gunicorn